requirements.txt
pytest==7.4.0
requests==2.31.0
requests-mock==1.11.0
allure-pytest==2.13.2
faker==19.6.2
pyyaml==6.0
